updated user.posts slice of state
